requirements.txt
python-dotenv
ollama
requests
openai
gradio